first crack at consolidation. nibread and nibwrite both use shared fileio routines.  still hashing out some things, but nibconv will be next.
git-svn-id: https://c64preservation.com/svn/nibtools/trunk@10 048eaad1-c872-4173-8c30-523bb29f53ec

+++ nibscan.c
@@ -14,7 +14,7 @@
 #include "nibtools.h"
 #include "version.h"
 
-BYTE start_track, end_track, track_inc;
+int start_track, end_track, track_inc;
 int imagetype, mode;
 int align, force_align;
 char bitrate_range[4] = { 43 * 2, 31 * 2, 25 * 2, 18 * 2 };
